+++ oereb_client/static/src/component/menu/menu.jsx
@@ -7,7 +7,7 @@ import {useDispatch, useSelector} from 'react-redux';
 
 import {setViewServices} from '../../reducer/accordion';
 import {showAvailability} from '../../reducer/availability';
-import {loadExtract, showError, showExtract, toggleBackground, toggleHighlight} from '../../reducer/extract';
+import {loadExtract, showError, showExtract} from '../../reducer/extract';
 import {updateHistory} from '../../reducer/history';
 import {hide, loadAt, show} from '../../reducer/map_query';
 import {enableSymbolZoom} from '../../reducer/symbol_zoom';
@@ -239,36 +239,6 @@ const OerebMenu = function () {
     dispatch(enableSymbolZoom(!symbolZoomEnabled));
   }
 
-  const realEstateText = (() => {
-    if (highlight) {
-      return (
-        <span><i className="bi bi-check-square"></i> {t('menu.settings.show_real_estate')}</span>
-      );
-    }
-    return (
-      <span><i className="bi bi-square"></i> {t('menu.settings.show_real_estate')}</span>
-    );
-  })();
-
-  const toggleRealEstate = function () {
-    dispatch(toggleHighlight());
-  }
-
-  const backgroundMapText = (() => {
-    if (background) {
-      return (
-        <span><i className="bi bi-check-square"></i> {t('menu.settings.show_background_map')}</span>
-      );
-    }
-    return (
-      <span><i className="bi bi-square"></i> {t('menu.settings.show_background_map')}</span>
-    );
-  })();
-
-  const toggleBackgroundMap = function () {
-    dispatch(toggleBackground());
-  }
-
   const historyElements = history.map((element, key) =>
     <li key={key}>
       <button className="dropdown-item"
@@ -312,16 +282,6 @@ const OerebMenu = function () {
             <li>
               <button className="dropdown-item" onClick={toggleSymbolZoom}>
                 {symbolZoomText}
-              </button>
-            </li>
-            <li>
-              <button className="dropdown-item" onClick={toggleRealEstate}>
-                {realEstateText}
-              </button>
-            </li>
-            <li>
-              <button className="dropdown-item" onClick={toggleBackgroundMap}>
-                {backgroundMapText}
               </button>
             </li>
           </ul>