+++ oereb_client/static/src/component/toggle_highlight/toggle_highlight.jsx
@@ -1,0 +1,38 @@
+import '../../oereb_client.scss';
+
+import React from 'react';
+import {useTranslation} from 'react-i18next';
+import { useDispatch, useSelector } from 'react-redux';
+import { toggleHighlight } from '../../reducer/extract';
+
+/**
+ * A simple button component to hide the real estate highlight.
+ */
+const OerebToggleHighlight = function () {
+  const {t} = useTranslation();
+  const highlight = useSelector((state) => state.extract).highlight;
+  const dispatch = useDispatch();
+
+  const toggle = () => {
+    dispatch(toggleHighlight());
+  };
+
+  let buttonClass;
+  if (highlight) {
+    buttonClass = "btn btn-outline-secondary";
+  }
+  else {
+    buttonClass = "btn btn-secondary";
+  }
+
+  return (
+    <button type="button"
+      className={buttonClass}
+      onClick={toggle}
+      title={t('extract.hide_highlight.title')}>
+      <i className="bi bi-bounding-box-circles"></i>
+    </button>
+  );
+};
+
+export default OerebToggleHighlight;